requirements.txt
aiofiles==23.1.0
blinker==1.5
click==8.1.6
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.14.4
hyperframe==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
priority==2.0.0
pyee==11.0.0
python-ffmpeg==2.0.4
quart==0.18.4
typing_extensions==4.7.1
Werkzeug==2.3.6
wsproto==1.2.0
